migrate rewrite proxy listener to handler
