+++ localstack/plugins.py
@@ -21,9 +21,10 @@ def register_partition_adjusting_proxy_listener():
     LOG.info(
         "Registering ArnPartitionRewriteListener to dynamically replace partitions in requests and responses."
     )
-    from localstack.services.generic_proxy import ArnPartitionRewriteListener, ProxyListener
+    from localstack.aws import handlers
+    from localstack.aws.handlers.partition_rewriter import ArnPartitionRewriteHandler
 
-    ProxyListener.DEFAULT_LISTENERS.append(ArnPartitionRewriteListener())
+    handlers.preprocess_request.append(ArnPartitionRewriteHandler())
 
 
 @hooks.on_infra_start()
